added test case for segmented modeler
